Changes:
Finished the configuration file for the beans with the hibernate and Database.
Added @Repository to the repositories.
Fixed the injection in the controllers.
Fixed the main class as well.

TO DO:
Fix the dependencies in the POM.xml because it's a total mess.
Fix the module-info.java file because it's a mess as well.